requirements.txt
numpy
onnx
onnxruntime
torch
torchvision
ultralytics
pycocotools
opencv-python
Pillow
tqdm
